% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/matrix_element_extra_parts.R
\name{clebsch_gordan2}
\alias{clebsch_gordan2}
\title{Clebsch-Gordan Coeffcient Alternate Version.}
\usage{
clebsch_gordan2(j1, j2, m1, m2, j, mj)
}
\arguments{
\item{j1}{A numeric. The total angular momentum for state 1.}

\item{j2}{A numeric. The total angular momentum for state 2.}

\item{m1}{A numeric. The magnetic quantum number for state 1.}

\item{m2}{A numeric. The magnetic quantum number for state 2.}

\item{j}{A numeric. The combined total angular momentum.}

\item{mj}{A numeric. The combined magnetic quantum number.}
}
\description{
\code{clebsch_gordan} calculates the Clebsch-Gordan coefficient
}
\details{
This function calculates the Clebsch-Gordan coeffcient for an arbitrary
selection of quantum numbers j_1, j_2, m_1, m_2, j, and m_j. The condition
that must be met is (m_1 + m_2) = m_j. The analytic form used to calculate
the Clebsch-Gordan coefficient can be found in Cornwell's Group Theory of
Physics.
}
\examples{
clebsch_gordan2(1, 1 / 2, 0, 1 / 2, 3 / 2, 1 / 2)
}

