% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/quantum_defect.R
\name{quantum_defect}
\alias{quantum_defect}
\title{Quantum Defect Calculation.}
\usage{
quantum_defect(n, l, j)
}
\arguments{
\item{n}{A numeric. The principle quantum number n.}

\item{l}{A numeric. The orbital angular momentum quantum number l.}

\item{j}{A numeric. The total angular momentum quantum number j.}
}
\description{
\code{quantum_defect} calculates the quantum defect for an arbitrary state
(n, l, j) for Rubidium-85
}
\details{
This function calculates the quantum defect for an arbitrary state
 (n, l, j) in Rubidium-85. Depending on the choices of l and j, the proper
 parameters will be selected and the quantum defect will be calculated
 according to equation (16.19) on pg. 351 of Rydberg Atoms by Gallagher.
 For example, given an arbitrary n, l = 1, and j = 1/2, the quantum defect
 for the state np_1/2 would be calculated.
}
\examples{
quantum_defect(32, 1, 1.5)
 quantum_defect(5, 0, 0.5)
}

