% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/radial_matrix_element.R
\name{radial_matrix_element}
\alias{radial_matrix_element}
\title{Radial Matrix Element.}
\usage{
radial_matrix_element(k, n1, n2, l1, l2, j1, j2)
}
\arguments{
\item{k}{A numeric. The power of r to be calculated over. To get a dipole
matrix element, k must be equal to 1.}

\item{n1}{A numeric. The principle quantum number of state 1.}

\item{n2}{A numeric. The principle quantum number of state 2.}

\item{l1}{A numeric. The orbital angular momentum number of state 1.}

\item{l2}{A numeric. The orbital angular momentum number of state 2.}

\item{j1}{A numeric. The total angular momentum number of state 1.}

\item{j2}{A numeric. The total angular momentum number of state 2.}
}
\description{
\code{radial_matrix_element} calculates the radial matrix element.
}
\details{
This function calculates the radial matrix element for two arbitrary states
(n1, l1, j1) and (n2, l2, j2). A Numerov algorithm is used to compute the
radial matrix elements as done in Appendix A of Zimmerman et al, PRA, 20, 2251
(1979). The scaling used in this function is \eqn{\xi = \sqrt r},
\eqn{\Psi = r^(3/4) R(r)} as done by Bhatti, Cromer, and Cooke, PRA, 24, 161
(1981).
}

