% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_crossing.R
\name{resonance}
\alias{resonance}
\title{Two-body Resonance Calculator.}
\usage{
resonance(Frame, AFrame, StateB, StateC)
}
\arguments{
\item{Frame}{A data frame. A data frame containing all of the in the Stark
map. Needs columns Field, Ecm, and state.}

\item{AFrame.}{A data frame. Needs columns Field and Ecm.}

\item{StateB.}{A character vector identifying state B.}

\item{StateC.}{A character vector identifying state C.}
}
\description{
\code{resonance} calculates the energy crossover of a pair of atoms in state
A and a pair state of atoms in state B and state C
}
\details{
This function calculates the energy \eqn{(2*E_A - (E_B + E_C))}. The function
requires a data frame with all of the states and their field values such that
Frame$Field gives the electric field values and Frame$Ecm gives the energy in
inverse centimeters (cm^-1). The second argument, AFrame, is a separate data
frame that already has the frame for state A selected. The other two inputs
should be strings that correspond to a the state column in Frame.
}

