% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_element_extra_parts.R
\name{sphere_mat_element}
\alias{sphere_mat_element}
\title{Spherical Matrix Element.}
\usage{
sphere_mat_element(l, lprime, ml)
}
\arguments{
\item{l}{A numeric. The orbital angular momentum of the starting state.}

\item{lprime}{A numeric. The orbital angular momentum of the final state.}

\item{ml}{A numeric. The magnetic momentum of both the starting and final
state. The m_l of each state must be equal.}
}
\description{
\code{sphere_mat_element} calculates the spherical matrix element
}
\details{
This function calculates the spherical matrix element for a given l, l',
and m_l. It finds the matrix element <l, m|cos(theta)|l', m> based on a
paper by Zimmerman et al, PRA 20, 2251 (1979).
}

