% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stark_map.R
\name{stark_matrix}
\alias{stark_matrix}
\title{Builds a Stark Matrix.}
\usage{
stark_matrix(n, l, j, mj)
}
\arguments{
\item{n}{A numeric. The vector of principle quantum numbers.}

\item{l}{A numeric. The vector of orbital angular momentum quantum numbers.}

\item{j}{A numeric. The vector of total angular momentum quantum numbers.}

\item{mj}{A numeric. The magnetic momentum quantum number.}
}
\description{
\code{stark_matrix} builds a Stark matrix.
}
\details{
This function builds a Stark matrix to be used in the creation of a Stark
map. The function accepts vectors n, l, and j that describe all of the
quantum numbers to be used in building the Stark matrix. These vectors can
be obtained using \code{\link{state_list}} and are the columns of its
result. Only one value of mj should be provided as the Stark matricies
are mj independent. The Stark matrix is treated as symmetric and is only
computed for the uper right triangle of the matrix. Copies of those
elements are made in to their symmetric counterparts on the lower left
trinagle of the matrix.
}

