% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stark_matrix_element.R
\name{stark_matrix_elem}
\alias{stark_matrix_elem}
\title{Stark Matrix Element.}
\usage{
stark_matrix_elem(n1, n2, l1, l2, j1, j2, mj1, mj2)
}
\arguments{
\item{n1}{A numeric. The principle quantum number of state 1.}

\item{n2}{A numeric. The principle quantum number of state 2.}

\item{l1}{A numeric. The orbital angular momentum number of state 1.}

\item{l2}{A numeric. The orbital angular momentum number of state 2.}

\item{j1}{A numeric. The total angular momentum number of state 1.}

\item{j2}{A numeric. The total angular momentum number of state 2.}

\item{mj1}{A numeric. The magnetic angular momentum number of state 1.}

\item{mj2}{A numeric. The magnetic angular momentum number of state 2.}
}
\description{
\code{stark_matrix_element} calculates the elements of a Stark matrix
}
\details{
This function calculates contribution to the energy due to the Stark effect
for an arbitrary set of states (n1, l1, j1, mj1) and (n2, l2, j2, mj2). The
element is determined based on equation 10 from Zimmerman et al., PRA, 20,
2251 (1979). For a non-zero matrix element, the conditions mj1 = mj2 and l2 =
l1 +/- 1 must be met.
}
