% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stark_radial_matrix_element.R
\name{stark_radial_mat_elem}
\alias{stark_radial_mat_elem}
\title{Calculate Radial Matrix Elements Involving Stark States.}
\usage{
stark_radial_mat_elem(stark_mat, initial_state, stark_state, field, n_min,
  n_max, ...)
}
\arguments{
\item{stark_mat}{A matrix that is produced using \code{\link{stark_matrix}}
that must correspond to n_min and m_max.}

\item{initial_state}{A character representing the initial state in the form
"n,l,j,m_j".}

\item{stark_state}{A character representing the state at zero field that
corresponds to the Stark state and has the form "n,l,j,m_j".}

\item{field}{A numeric representing the field strength at which the Stark
state is evaluated.}

\item{n_min}{A numeric giving the minimum n used in determining the size
of the Stark matrix.}

\item{n_max}{A numeric giving the maximum n used in determining the size
of the Stark matrix.}

\item{...}{Additional parameters to be sent to other functions.}
}
\description{
\code{stark_radial_mat_elem} calculates the radial matrix element between
a given initial state and a Stark state.
}
\details{
This function calculates the radial matrix elements between a given initial
state and a Stark state at some arbitrary field F. The Stark state that is
given as input is the state that a particular line would correspond with at
zero field. The Stark matrix required for the function is produced by
\code{\link{stark_matrix}} and must correspond to the provided n_min and
n_max.
}

