% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stark_map.R
\name{tidy_stark_energy}
\alias{tidy_stark_energy}
\title{Tidy Stark Energy Data Frame.}
\usage{
tidy_stark_energy(zero_frame, stark_energy, field_min, field_max, step_size)
}
\arguments{
\item{zero_frame}{A data frame that contains the zero electric field
energies and state information.}

\item{stark_energy}{A matrix that contains the eigen energies for the
states in the columns and the field levels in the rows.}

\item{field_min}{A numeric that sets the mininum of the electric field.}

\item{field_max}{A numeric that sets the maximum of the electric field.}

\item{step_size}{A numeric that sets the step size of the electric field.}
}
\description{
\code{tidy_stark_energy} creates a tidy data frame of the Stark energy
}
\details{
This function creates a tidy data frame of the Stark energy. The zero energy
data frame is provided by the function \code{\link{zero_field_energy_df}}.
The Stark energy matrix is provided by the function
\code{\link{stark_eigen}}. The field_min, field_max, and step_size should
be set to same values as when \code{stark_eigen} was run.
}

