% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/matrix_element_extra_parts.R
\name{wigner_3j}
\alias{wigner_3j}
\title{Wigner 3j Symbol.}
\usage{
wigner_3j(j1, j2, j3, m1, m2, m3)
}
\arguments{
\item{j1}{A numeric.}

\item{j2}{A numeric.}

\item{j3}{A numeric.}

\item{m1}{A numeric.}

\item{m2}{A numeric.}

\item{m3}{A numeric.}
}
\description{
\code{wigner_3j} calculates the Wigner 3j symbol using the Clebsch-Gordan
coefficient calculated in \code{\link{clebsch_gordan}}.
}
\details{
This function calculates the Wigner 3j symbol using the Clebsch-Gordan
coefficient calculated in \code{\link{clebsch_gordan}}. The Wigner 3j symbol
can be written as follows:
\tabular{ccc}{
j1 \tab j2 \tab j3\cr
m1 \tab m2 \tab m3
}
}

