% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_crossing.R
\name{zero_cross}
\alias{zero_cross}
\title{Zero Crossing.}
\usage{
zero_cross(x, y)
}
\arguments{
\item{x}{A numeric. The independent variable. An associated position or time
vector to y.}

\item{y}{A numeric. The dependent variable. Typically a signal or function
of the independent variable x.}
}
\description{
\code{zero_cross} calculates the zero crossings for a pair of vectors x
and y.
}
\details{
This function calculates the zero crossing for vectors x and y. The vector y
is the dependent variable and the vector x is the indepdenent variable.
}

