% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stark_map.R
\name{zero_field_energy_mat}
\alias{zero_field_energy_mat}
\title{Construct Zero Field Energy Matrix.}
\usage{
zero_field_energy_mat(n, l, j)
}
\arguments{
\item{n}{A numeric. A vector that contains a series of principle quantum
numbers.}

\item{l}{A numeric. A vector that contains a series of orbital angular
momentum quantum numbers.}

\item{j}{A numeric. A vector that contains a series of total angular
momentum quantum numbers.}
}
\description{
\code{zero_field_energy_mat} constructs a matrix with the zero field
energies on the diagonal.
}
\details{
This function takes vectors for n, l, and j and returns a matrix with the
energies for those n, l, and j states on the diagonal. To create the
vectors for n, l, and j, \code{\link{state_list}} should be used. This will
create a matrix with all the combinations of n, l, and j needed to create
the Stark matrix. The energy is determined by:
\deqn{E = -1 / (2* (n - \delta(n, l, j))^2)}
where \eqn{\delta(n, l, j)} is determined by \code{\link{quantum_defect}}.
}

