% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stark_map.R
\name{stark_eigen}
\alias{stark_eigen}
\title{Eigenvalues of the Stark Matrix.}
\usage{
stark_eigen(stark_matrix, zero_field_mat, field_min, field_max, step_size)
}
\arguments{
\item{stark_matrix}{A matrix created using the function
\code{\link{stark_matrix}}.}

\item{zero_field_mat}{A matrix created using the function
\code{\link{zero_field_energy_mat}}.}

\item{field_min}{A numeric that sets the mininum of the electric field.}

\item{field_max}{A numeric that sets the maximum of the electric field.}

\item{step_size}{A numeric that sets the step size of the electric field.}
}
\description{
\code{stark_eigen} calculates the eigenvalues of the Stark matrix for a
sequence of electric field sizes.
}
\details{
This function calculates the eigenvalues of the sum of the the zero electric
field matrix and the Stark matrix times the electric field. The eigenvalues
are output as a matrix with the different states as the columns and each
field step as the rows. The columns are in descending order and the rows
are in the initial state order of the zero electric field matrix and Stark
matrix.
}
